+++ babel.config.js
@@ -1,7 +1,11 @@
 const path = require('path')
+const emotionDevPreset = require('babel-preset-emotion-dev')
 
-let needsBabelPluginEmotion = filename =>
-  filename.includes(path.join('.test.js'))
+let needsBabelPluginEmotion = filename => /\.test\.js$/.test(filename)
+
+let isTestFile = filename =>
+  /\.test\.js$/.test(filename) ||
+  filename.includes(`${path.sep}__tests__${path.sep}`)
 
 module.exports = api => {
   api.cache(true)
@@ -14,14 +18,32 @@ module.exports = api => {
           ((!filename.includes('no-babel') &&
             needsBabelPluginEmotion(filename)) ||
             filename.includes(path.join('__tests__', 'babel'))),
-        plugins: ['@emotion/babel-plugin']
+        presets: [[emotionDevPreset, { useEmotionPlugin: true }]]
       },
       {
         test: filename =>
           filename &&
           filename.includes('source-map') &&
           needsBabelPluginEmotion(filename),
-        plugins: [['@emotion/babel-plugin', { sourceMap: true }]]
+        presets: [
+          [emotionDevPreset, { useEmotionPlugin: true, sourceMap: true }]
+        ]
+      },
+      {
+        test: filename =>
+          filename &&
+          isTestFile(filename) &&
+          filename.includes('automatic-runtime'),
+        presets: [[emotionDevPreset, { runtime: 'automatic' }]]
+      },
+      {
+        test: filename =>
+          filename &&
+          isTestFile(filename) &&
+          filename.includes('automatic-dev-runtime'),
+        presets: [
+          [emotionDevPreset, { runtime: 'automatic', development: true }]
+        ]
       }
     ]
   }